refactor: removed enum
